+++ tests/testthat/test_findCovs.R
@@ -28,6 +28,11 @@ test_that("with by of length 2",{
 
 
 ## 
-test_that("deprecated cols.id",
-          expect_warning(findCovs(data.frame(x=1),cols.id="w"))
-          )
+test_that("deprecated cols.id",{
+    expect_warning(
+        findCovs(data.frame(x=rep(1:2,each=2),y=c(rep(c("a"),2),"v","w")),cols.id="x")
+    )
+    expect_error(
+        findCovs(data.frame(x=rep(1:2,each=2),y=c(rep(c("a"),2),"v","w")),cols.id="x",by="x")
+    )
+})